feat: checkButton 구현 및 success 페이지 구현
