+++ src/App.css
@@ -17,30 +17,3 @@ html body {
 body {
   padding: 20px;
 }
-
-.App {
-  max-width: 300px;
-  width: 100%;
-  justify-self: center;
-  display: flex;
-  flex-direction: column;
-  padding: 50px 0;
-}
-
-.card-input {
-  width: 100%;
-  justify-self: center;
-  display: flex;
-  flex-direction: column-reverse;
-}
-
-.checkButton-container {
-  padding: 18px 0 8px 0;
-}
-
-.check-button {
-  background-color: var(--color-black);
-  color: var(--color-white);
-  width: 100%;
-  height: 52px;
-}
