+++ src/components/CheckButton.tsx
@@ -1,7 +1,17 @@
-export default function CheckButton() {
+interface CheckButtonProps {
+  className: string;
+  onSubmit: () => void;
+  text: string;
+}
+
+export default function CheckButton({
+  className,
+  onSubmit,
+  text,
+}: CheckButtonProps) {
   return (
-    <button className="check-button">
-      <span>확인</span>
+    <button className={className} onClick={onSubmit}>
+      {text}
     </button>
   );
 }
